% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIDBracket.R
\name{data_gen}
\alias{data_gen}
\title{Generate simulation datasets.}
\usage{
data_gen(N, case = c("case1", "case2", "S_t"), att = c(0, 2, 3, 1))
}
\arguments{
\item{N}{Sample size.}

\item{case}{Take values "case1", "case2", "S_t", used to indicate three simulation scenarios in Ye et al., (2020).}

\item{att}{The true treatment effects for the treated group respectively at t=1,2,3,4. The first element needs to be 0 since this corresponds to the pre-treatment period. Default is c(0,2,3,1).}
}
\description{
Generate simulation datasets.
}
\examples{
data_gen(1000,"case1")

}
\references{
Ting Ye, Luke Keele, Raiden Hasegawa, Dylan S. Small (2020). A Negative Correlation Strategy for Bracketing in Difference-in-Differences with Application to the Effect of Voter Identification Laws on Voter Turnout.
}
