% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_S_t}
\alias{df_S_t}
\title{Simulated datasets for Section 6.2}
\format{An object of class \code{data.frame} with 4000000 rows and 6 columns.}
\usage{
data(df_S_t)
}
\description{
1000 long-form data frames generated from the same data generating process. Each with six columns (g, t, Y, S, simu, case),
where column g is the group indicator, column t is the time indicator,
column Y is the observed outcome, column S is the observed time-varying covariate, column simu indicates the simulation repetition,
column case is the case indicator.
Specifically, g takes values trt, a, b where a and b represent
the two control groups; t takes values 1, 2, 3, 4,
where the treated group adopts the treatment between t=1 and t=2.
}
\references{
Ting Ye, Luke Keele, Raiden Hasegawa, Dylan S. Small. A Negative Correlation Strategy for Bracketing in Difference-in-Differences with Application to the Effect of Voter Identification Laws on Voter Turnout.
}
\keyword{datasets}
