% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIDBracket.R
\name{did_list}
\alias{did_list}
\title{Calculate the DID parameters used to construct the brackets.}
\usage{
did_list(df, l)
}
\arguments{
\item{df}{A long-form data frame with three columns (g, t, Y), where column g is the group indicator, column t is the time indicator, column Y is the observed outcome. Specifically, g takes values trt, a, b, where a and b represent the two control groups; t takes values 1, 2, ..., where the treated group adopts the treatment between t=1 and t=2.}

\item{l}{Estimate the treatment effect for the treated group l time periods after adopting the treatment.}
}
\description{
Calculate the DID parameters used to construct the brackets.
}
\examples{

df<-data_gen(1000,"case1")
did_list(df,l=2)

}
